pyproject.toml
[tool.poetry]
name = "personal-assistant-petproject"
version = "0.1.0"
description = ""
authors = ["Vchslv-Zkrhv <vchslv.zkhrv@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
aiogram = "^3.3.0"
python-telegram-bot = "^20.7"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
